requirements.txt
Flask==3.0.3
pymysql==1.1.1
requests==2.32.3
python_dotenv==1.0.1
Flask-Session==0.8.0
Flask-RESTful==0.3.9
gunicorn==21.2.0

